requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
psycopg2-binary
python-multipart==0.0.9
pgvector==0.3.6

yt-dlp
Pillow
cryptography
gtts
reportlab
pyyaml==6.0.2
